pyproject.toml
[project]
name = "rtfm-rag"
version = "0.1.0"
description = "rtfm-rag"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.7",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.12",
    "html2text>=2025.4.15",
    "openai>=1.83.0",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "result>=0.17.0",
]

[dependency-groups]
dev = [
    "dash>=3.0.4",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "umap-learn>=0.5.7",
]
